Adde lib/apps
